initial add of filtered packet channels for received packets
